Ocean_Lens/requirements.txt
kornia==0.7.3
numpy==1.24.4
opencv_python==4.10.0.84
pandas==2.0.3
Pillow==11.0.0
scipy==1.14.1
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
